promoting com.google.fonts/check/family_has_same_vertical_metrics to universal
(#2468)
