+++ Lib/fontbakery/profiles/universal.py
@@ -36,7 +36,7 @@ UNIVERSAL_PROFILE_CHECKS = \
   'com.google.fonts/check/valid_glyphnames',
   'com.google.fonts/check/unique_glyphnames',
 #  'com.google.fonts/check/glyphnames_max_length',
-
+  'com.google.fonts/check/family_has_same_vertical_metrics',
 ]
 
 @check(
@@ -942,6 +942,50 @@ def com_google_fonts_check_ttx_roundtrip(font):
   # and then we need to cleanup our mess...
   if os.path.exists(xml_file):
     os.remove(xml_file)
+
+
+@check(
+  id = 'com.google.fonts/check/family_has_same_vertical_metrics',
+  rationale="""
+  We want all fonts within a family to have the same vertical metrics so
+  their line spacing is consistent across the family.
+  """
+)
+def com_google_fonts_check_family_has_same_vertical_metrics(ttFonts):
+  """Each font in a family must have the same vertical metrics values."""
+  failed = []
+  vmetrics = {
+    "sTypoAscender": {},
+    "sTypoDescender": {},
+    "sTypoLineGap": {},
+    "usWinAscent": {},
+    "usWinDescent": {},
+    "ascent": {},
+    "descent": {}
+  }
+
+  for ttfont in ttFonts:
+      full_font_name = ttfont['name'].getName(4, 3, 1, 1033).toUnicode()
+      vmetrics['sTypoAscender'][full_font_name] = ttfont['OS/2'].sTypoAscender
+      vmetrics['sTypoDescender'][full_font_name] = ttfont['OS/2'].sTypoDescender
+      vmetrics['sTypoLineGap'][full_font_name] = ttfont['OS/2'].sTypoLineGap
+      vmetrics['usWinAscent'][full_font_name] = ttfont['OS/2'].usWinAscent
+      vmetrics['usWinDescent'][full_font_name] = ttfont['OS/2'].usWinDescent
+      vmetrics['ascent'][full_font_name] = ttfont['hhea'].ascent
+      vmetrics['descent'][full_font_name] = ttfont['hhea'].descent
+
+  for k, v in vmetrics.items():
+      metric_vals = set(vmetrics[k].values())
+      if len(metric_vals) != 1:
+          failed.append(k)
+
+  if failed:
+      for k in failed:
+        s = ["{}: {}".format(k, v) for k, v in vmetrics[k].items()]
+        yield FAIL, ("{} is not the same across the family:\n:"
+                     "{}".format(k, "\n".join(s)))
+  else:
+      yield PASS, "Vertical metrics are the same across the family"
 
 
 profile.auto_register(globals())
