+++ tests/profiles/universal_test.py
@@ -967,3 +967,14 @@ def test_check_os2_metrics_match_hhea(mada_ttFonts):
   status, message = list(check(ttFont))[-1]
   assert status == FAIL and message.code == "descender"
 
+
+def test_check_family_has_same_vertical_metrics(montserrat_ttFonts):
+  from fontbakery.profiles.universal import com_google_fonts_check_family_has_same_vertical_metrics as check
+  print("Test pass with multiple good fonts...")
+  status, message = list(check(montserrat_ttFonts))[-1]
+  assert status == PASS
+
+  print("Test fail with one bad font that has one different vertical metric val...")
+  montserrat_ttFonts[0]['OS/2'].usWinAscent = 4000
+  status, message = list(check(montserrat_ttFonts))[-1]
+  assert status == FAIL
